third_party/blink/web_tests/wpt_internal/attribution-reporting/source-registration.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<meta name=variant content="?method=a">
<meta name=variant content="?method=a&eligible">
<meta name=variant content="?method=fetch&eligible=event-source">
<meta name=variant content="?method=fetch&eligible=event-source&cross-origin">
<meta name=variant content="?method=img">
<meta name=variant content="?method=img&eligible">
<meta name=variant content="?method=open">
<meta name=variant content="?method=open&eligible">
<meta name=variant content="?method=script">
<meta name=variant content="?method=script&eligible">
<meta name=variant content="?method=xhr&eligible=event-source">
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<body>
<script>
attribution_reporting_promise_test(async t => {
  const reportingOrigin = getDefaultReportingOrigin();
  const source = {
    source_event_id: generateSourceEventId(),
    destination: `https://{{host}}`,
    debug_reporting: true,
  };
  const sourceType = await registerAttributionSrc(t, {
    source,
    method: 'variant',
    reportingOrigin,
    cookie: attributionDebugCookie,
  });

  console.log('source-event-id', source.source_event_id);

  await waitForSourceToBeRegistered(source.source_event_id);
  await registerAttributionSrc(t, {
    trigger: {
      event_trigger_data: [{}],
    },
    reportingOrigin,
  });

  const payload = await pollEventLevelReports(reportingOrigin);
  console.log('event-level-reports', JSON.stringify(payload));
  assert_equals(payload.reports.length, 1);

  const report = JSON.parse(payload.reports[0].body);
  assert_equals(report.source_event_id, source.source_event_id);
  assert_equals(report.source_type, sourceType);

  const debugReportPayload = await pollVerboseDebugReports(reportingOrigin);
  console.log('verbose-debug-reports', JSON.stringify(debugReportPayload));
  assert_equals(debugReportPayload.reports.length, 1);
  const debugReport = JSON.parse(debugReportPayload.reports[0].body);
  assert_equals(debugReport.length, 1);
  assert_equals(debugReport[0].type, 'source-success');
}, 'Source registration succeeds.');
</script>
